requirements.txt
selenium
MemePy
pyquery
pyautogui
python-twitter
pyyaml
tweepy